Refactor command-handler into handlers folder
