+++ index.ts
@@ -1,10 +1,9 @@
 // https://github.com/discordjs/voice/tree/main/examples/music-bot
 
-import { Client, GuildMember, Intents, Snowflake } from 'discord.js';
-import { joinVoiceChannel, VoiceConnectionStatus, entersState, AudioPlayerStatus, AudioResource } from '@discordjs/voice';
+import { Client, Intents, Snowflake } from 'discord.js';
 import { token } from './config.json';
 import { MusicSubscription } from './src/music/subscription';
-import { handleCommandByName } from './src/command-handler';
+import { handleCommandByName } from './src/handlers';
 
 const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_VOICE_STATES]})
 const subscriptions = new Map<Snowflake, MusicSubscription>();
