+++ src/util.ts
@@ -1,0 +1,11 @@
+import { Snowflake } from "discord.js";
+import { MusicSubscription } from "./music/subscription";
+
+export const getSubscription = (subscriptions: Map<Snowflake, MusicSubscription>, guildId: string | null) => {
+    if (guildId === null) {
+      return undefined;
+    }
+  
+    return subscriptions.get(guildId);
+  }
+
